Move GasUnit tests into it's submodule
