+++ crates/module-system/sov-modules-api/src/gas/tests/mod.rs
@@ -1,3 +1,2 @@
 mod charge_gas;
-mod json_schema;
 mod metered_utils;
